+++ app/src/main/java/com/geridea/trentastico/model/ExtraCoursesList.kt
@@ -11,7 +11,7 @@ import java.util.*
 class ExtraCoursesList : ArrayList<ExtraCourse>() {
 
     fun isAnExtraLesson(lesson: LessonSchedule): Boolean =
-            this.any { it.lessonTypeId.toLong() == lesson.lessonTypeId }
+            this.any { it.lessonTypeId == lesson.lessonTypeId }
 
     fun toJSON(): JSONArray {
         val jsonArray = JSONArray()
